refactor(api)!: Refactor secret deletion in compositions
Deleting secrets is a composition, because it refers/uses both secrets and
infrastructure services components, so it is clearest to refactor it
separately.

Resolves #3577.
