+++ compositions/secrets-routes/src/test/kotlin/routes/DeleteProductSecretIntegrationTest.kt
@@ -33,41 +33,43 @@ import java.util.EnumSet
 
 import org.eclipse.apoapsis.ortserver.compositions.secretsroutes.SecretsRoutesIntegrationTest
 import org.eclipse.apoapsis.ortserver.model.CredentialsType
-import org.eclipse.apoapsis.ortserver.model.OrganizationId
+import org.eclipse.apoapsis.ortserver.model.ProductId
 import org.eclipse.apoapsis.ortserver.model.repositories.SecretRepository
 import org.eclipse.apoapsis.ortserver.secrets.Path
 import org.eclipse.apoapsis.ortserver.secrets.SecretsProviderFactoryForTesting
 import org.eclipse.apoapsis.ortserver.shared.apimodel.ErrorResponse
 
-class DeleteSecretByOrganizationIdAndNameIntegrationTest : SecretsRoutesIntegrationTest({
-    var orgId = 0L
+class DeleteProductSecretIntegrationTest : SecretsRoutesIntegrationTest({
+    var prodId = 0L
+
+    val secretErrorPath = "error-path"
 
     beforeEach {
-        orgId = dbExtension.fixtures.organization.id
+        prodId = dbExtension.fixtures.product.id
     }
 
-    "DeleteSecretByOrganizationIdAndName" should {
+    "DeleteProductSecret" should {
         "delete a secret" {
             secretsRoutesTestApplication { client ->
-                val secret = secretRepository.createOrganizationSecret(orgId)
+                val secret = secretRepository.createProductSecret(prodId)
 
-                client.delete("/organizations/$orgId/secrets/${secret.name}") shouldHaveStatus
+                client.delete("/products/$prodId/secrets/${secret.name}") shouldHaveStatus
                         HttpStatusCode.NoContent
 
-                secretRepository.listForId(OrganizationId(orgId)).data shouldBe emptyList()
+                secretRepository.listForId(ProductId(prodId)).data shouldBe emptyList()
 
                 val provider = SecretsProviderFactoryForTesting.instance()
                 provider.readSecret(Path(secret.path)) should beNull()
             }
         }
 
-        "respond with 'Conflict' when secret is in use" {
+        "respond with Conflict when secret is in use" {
             secretsRoutesTestApplication { client ->
-                val userSecret = secretRepository.createOrganizationSecret(orgId, path = "user", name = "user").name
-                val passSecret = secretRepository.createOrganizationSecret(orgId, path = "pass", name = "pass").name
+                val userSecret = secretRepository.createProductSecret(prodId, path = "user", name = "user").name
+                val passSecret = secretRepository.createProductSecret(prodId, path = "pass", name = "pass").name
 
                 val service = infrastructureServiceService.createForId(
-                    OrganizationId(orgId),
+                    ProductId(prodId),
                     name = "testService",
                     url = "http://repo1.example.org/obsolete",
                     description = "good bye, cruel world",
@@ -76,7 +78,7 @@ class DeleteSecretByOrganizationIdAndNameIntegrationTest : SecretsRoutesIntegrat
                     credentialsTypes = EnumSet.of(CredentialsType.NETRC_FILE)
                 )
 
-                val response = client.delete("/organizations/$orgId/secrets/$userSecret")
+                val response = client.delete("/products/$prodId/secrets/$userSecret")
                 response shouldHaveStatus HttpStatusCode.Conflict
 
                 val body = response.body<ErrorResponse>()
@@ -87,20 +89,20 @@ class DeleteSecretByOrganizationIdAndNameIntegrationTest : SecretsRoutesIntegrat
 
         "handle a failure from the SecretStorage" {
             secretsRoutesTestApplication { client ->
-                val secret = secretRepository.createOrganizationSecret(orgId, path = secretErrorPath)
+                val secret = secretRepository.createProductSecret(prodId, path = secretErrorPath)
 
-                client.delete("/organizations/$orgId/secrets/${secret.name}") shouldHaveStatus
+                client.delete("/products/$prodId/secrets/${secret.name}") shouldHaveStatus
                         HttpStatusCode.InternalServerError
 
-                secretRepository.getByIdAndName(OrganizationId(orgId), secret.name) shouldBe secret
+                secretRepository.getByIdAndName(ProductId(prodId), secret.name) shouldBe secret
             }
         }
     }
 })
 
-fun SecretRepository.createOrganizationSecret(
-    orgId: Long,
+fun SecretRepository.createProductSecret(
+    prodId: Long,
     path: String = "path",
     name: String = "name",
     description: String = "description"
-) = create(path, name, description, OrganizationId(orgId))
+) = create(path, name, description, ProductId(prodId))
